chore: removes extra accordion examples
