+++ screener.config.js
@@ -8,7 +8,6 @@ LICENSE file in the root directory of this source tree.
 /* eslint-env node */
 /* eslint-disable flowtype/require-valid-file-annotation */
 
-const Accordion = require('./src/accordion/examples-list');
 const Button = require('./src/button/examples-list');
 const ButtonGroup = require('./src/button-group/examples-list');
 const Breadcrumbs = require('./src/breadcrumbs/examples-list');
@@ -35,7 +34,6 @@ module.exports = {
   resolution: '1024x768',
   baseBranch: 'master',
   includeRules: [
-    new RegExp(`${Accordion.ACCORDION_EXAMPLE}$`),
     new RegExp(Button.BUTTON),
     new RegExp(Button.BUTTON_COMPACT_WITH_ENHANCERS),
     new RegExp(Button.BUTTON_WITH_ENHANCERS),
