--- src/accordion/examples-list.js
@@ -1,14 +1,0 @@
-/*
-Copyright (c) 2018 Uber Technologies, Inc.
-
-This source code is licensed under the MIT license found in the
-LICENSE file in the root directory of this source tree.
-*/
-/* eslint-env node */
-/* eslint-disable flowtype/require-valid-file-annotation */
-
-module.exports = {
-  ACCORDION_EXAMPLE: 'Accordion',
-  STYLE_PROPS_OVERRIDES: 'Panel style and props overrides',
-  SINGLE_STATEFUL_PANEL: 'Single stateful panel',
-};
